+++ src/policies/envs/futures_env_v3.py
@@ -67,19 +67,19 @@ class FuturesEnvV3(gym.Env):
         # RL config
         self.max_steps = config.max_steps
         self.action_space: spaces.Box = spaces.Box(
-            low=-config.max_volume, high=config.max_volume, shape=(1,), dtype=np.int64)
+            low=-config.max_volume, high=config.max_volume, shape=(1,), dtype=np.int32)
 
         self.factor_length = 50
         self.observation_space: spaces.Dict = spaces.Dict({
-            # "last_volume": spaces.Box(low=-config.max_volume, high=config.max_volume, shape=(1,), dtype=np.int64),
-            "last_price": spaces.Box(low=0, high=1e10, shape=(1,), dtype=np.float64),
-            # "hour": spaces.Box(low=0, high=23, shape=(1,), dtype=np.int64),
-            # "minute": spaces.Box(low=0, high=59, shape=(1,), dtype=np.int64),
-            Interval.ONE_SEC.value: spaces.Box(low=0, high=1e10, shape=(self.data_length[Interval.ONE_SEC.value], 5), dtype=np.float64),
-            # Interval.ONE_MIN.value: spaces.Box(low=0, high=1e10, shape=(self.data_length[Interval.ONE_MIN.value], 5), dtype=np.float64),
-            # Interval.THIRTY_MIN.value: spaces.Box(low=0, high=1e10, shape=(self.data_length[Interval.THIRTY_MIN.value], 5), dtype=np.float64),
-            "macd_bar": spaces.Box(low=-np.inf, high=np.inf, shape=(self.factor_length, ), dtype=np.float64),
-            "rsi": spaces.Box(low=-np.inf, high=np.inf, shape=(self.factor_length,), dtype=np.float64),
+            # "last_volume": spaces.Box(low=-config.max_volume, high=config.max_volume, shape=(1,), dtype=np.int32),
+            "last_price": spaces.Box(low=0, high=1e10, shape=(1,), dtype=np.float32),
+            # "hour": spaces.Box(low=0, high=23, shape=(1,), dtype=np.int32),
+            # "minute": spaces.Box(low=0, high=59, shape=(1,), dtype=np.int32),
+            Interval.ONE_SEC.value: spaces.Box(low=0, high=1e10, shape=(self.data_length[Interval.ONE_SEC.value], 5), dtype=np.float32),
+            # Interval.ONE_MIN.value: spaces.Box(low=0, high=1e10, shape=(self.data_length[Interval.ONE_MIN.value], 5), dtype=np.float32),
+            # Interval.THIRTY_MIN.value: spaces.Box(low=0, high=1e10, shape=(self.data_length[Interval.THIRTY_MIN.value], 5), dtype=np.float32),
+            "macd_bar": spaces.Box(low=-np.inf, high=np.inf, shape=(self.factor_length, ), dtype=np.float32),
+            "rsi": spaces.Box(low=-np.inf, high=np.inf, shape=(self.factor_length,), dtype=np.float32),
         })
 
     def _set_account(self, auth, backtest, init_balance, live_market=None, live_account=None):
@@ -133,14 +133,14 @@ class FuturesEnvV3(gym.Env):
             self.api.wait_update()
             if self.api.is_changing(self.bar_1s.iloc[-1], "datetime"):
 
-                bar_1s = self.bar_1s[self.OHLCV].iloc[-self.data_length[Interval.ONE_SEC.value]:].to_numpy(dtype=np.float64)
+                bar_1s = self.bar_1s[self.OHLCV].iloc[-self.data_length[Interval.ONE_SEC.value]:].to_numpy(dtype=np.float32)
 
                 state = dict({
-                    # "last_volume": np.array([self.last_volume], dtype=np.int64),
-                    "last_price": np.array([self.instrument_quote.last_price], dtype=np.float64),
+                    # "last_volume": np.array([self.last_volume], dtype=np.int32),
+                    "last_price": np.array([self.instrument_quote.last_price], dtype=np.float32),
                     Interval.ONE_SEC.value: bar_1s,
-                    "rsi": np.array(self.factors.rsi(self.bar_1s[-self.factor_length:], n=30)[-self.factor_length:], dtype=np.float64),
-                    "macd_bar": np.array(self.factors.macd_bar(self.bar_1s.iloc[-self.factor_length:], short=60, long=120, m=30), dtype=np.float64),
+                    "rsi": np.array(self.factors.rsi(self.bar_1s[-self.factor_length:], n=30)[-self.factor_length:], dtype=np.float32),
+                    "macd_bar": np.array(self.factors.macd_bar(self.bar_1s.iloc[-self.factor_length:], short=60, long=120, m=30), dtype=np.float32),
                 })
                 if np.isnan(bar_1s).any():
                     self.api.wait_update()
